Weapons and Toolbar
Weapons now shoot when in the player's hand. An enemy example is in the planet scene. Pickups exist and can place items into the players inventory that the player can move to their toolbar. the toolbar is navigated using the scroll wheel and should highlight active items red.
KNOWN BUGS:
If the player places a weapon into an active toolbar slot and clicks, the game will throw a null reference exception or use the previous item
Plants are scaled for the rack controller, so they appear minuscule in the player's hand